move first_set to grammar_table
